requirements.txt
pytorch==1.13.1 
torchvision==0.14.1 
torchaudio==0.13.1 
efficientnet_pytorch      
GitPython
pandas
tqdm
pyyaml
wandb
torch_optimizer
tensorboard